+++ packages/site-demo/content/product/components/thumbnail/vue/demo4.vue
@@ -1,0 +1,8 @@
+<template>
+    <Thumbnail image="https://picsum.photos/id/1/128/128" alt="Squared" size="xl" variant="squared" />
+    <Thumbnail image="https://picsum.photos/id/1/128/128" alt="Rounded" size="xl" variant="rounded" />
+</template>
+
+<script setup lang="ts">
+import { Thumbnail } from '@lumx/vue';
+</script>