+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Create the `BadgeWrapper` component
     -   Create the `Link` component
     -   Create the `Divider` component
+    -   Create the `Thumbnail` component
 
 ### Changed
 
@@ -24,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Moved `BadgeWrapper` from `@lumx/react`
     -   Moved `Link` from `@lumx/react`
     -   Moved `Divider` from `@lumx/react`
+    -   Moved `Thumbnail` from `@lumx/react`
     -   Uniformize documentation of `disabled` (as an alias to `isDisabled`) props across core components
 
 ## [4.3.1][] - 2026-02-11